refactor the code to make it clean
